+++ s/comms/json-rpc.ts
@@ -18,10 +18,21 @@ export namespace JsonRpc {
 		| Failure
 	)
 
+	export type Serializable = (
+		| void
+		| null
+		| undefined
+		| boolean
+		| number
+		| string
+		| Serializable[]
+		| {[key: string]: Serializable}
+	)
+
 	/////////////////////////////////////////////////////////////
 
 	export type Id = number | string | null
-	export type Params = any[] | {}
+	export type Params = Serializable[] | Record<string, Serializable>
 
 	export type Notification<P extends Params> = {
 		jsonrpc: string