+++ s/index.ts
@@ -7,13 +7,16 @@ export * from "./comms/json-rpc.js"
 export * from "./comms/respond.js"
 
 export * from "./core/errors.js"
-export * from "./core/expose.js"
+export * from "./core/endpoint.js"
 export * from "./core/remote-proxy.js"
 export * from "./core/remote.js"
 export * from "./core/types.js"
 
 export * from "./transports/http/client.js"
+export * from "./transports/post-message/post-messenger.js"
 export * from "./transports/websocket/client.js"
+export * from "./transports/utils/bidirectional.js"
+export * from "./transports/utils/response-waiter.js"
 
 export * from "./tools/logging/logger.js"
 export * from "./tools/logging/coloring.js"
@@ -21,4 +24,5 @@ export * from "./tools/logging/pretty-logger.js"
 export * from "./tools/logging/disabled-logger.js"
 export * from "./tools/deadline.js"
 export * from "./tools/errstring.js"
+export * from "./tools/simple-headers.js"