+++ s/transports/http/node-utils/endpoint-listener.ts
@@ -2,33 +2,37 @@
 import {RequestListener} from "http"
 
 import {readStream} from "./read-stream.js"
-import {Endpoint} from "../../../core/types.js"
 import {JsonRpc} from "../../../comms/json-rpc.js"
+import {ipAddress} from "../../../tools/ip-address.js"
+import {Endpoint, ServerMeta} from "../../../core/types.js"
+import {simplifyHeaders} from "../../../tools/simple-headers.js"
+import {defaults} from "../../defaults.js"
 
 export type EndpointListenerOptions = {
-	maxPayloadSize?: number
+	timeout?: number
+	maxRequestBytes?: number
 	onError?: (error: any) => void
 }
 
 export function makeEndpointListener(
-		endpoint: Endpoint,
+		makeEndpoint: (meta: ServerMeta) => Endpoint,
 		options: EndpointListenerOptions = {},
 	): RequestListener {
 
 	const {
-		maxPayloadSize = 10_000_000,
+		maxRequestBytes = defaults.maxRequestBytes,
 		onError = () => {},
 	} = options
 
 	return async(req, res) => {
 		try {
-			const {headers} = req
-			const body = await readStream(req, maxPayloadSize)
+			const body = await readStream(req, maxRequestBytes)
 			const requestish = JSON.parse(body) as JsonRpc.Requestish
-
-			const execute = async(request: JsonRpc.Request) => {
-				return await endpoint(request, {headers})
-			}
+			const endpoint = makeEndpoint({
+				req,
+				ip: ipAddress(req),
+				headers: simplifyHeaders(req.headers),
+			})
 
 			const send = (respondish: null | JsonRpc.Respondish) => {
 				res.statusCode = 200
@@ -37,7 +41,7 @@ export function makeEndpointListener(
 			}
 
 			if (Array.isArray(requestish)) {
-				const responses = (await Promise.all(requestish.map(execute)))
+				const responses = (await Promise.all(requestish.map(endpoint)))
 					.filter(r => !!r)
 				send(
 					(responses.length > 0)
@@ -46,7 +50,7 @@ export function makeEndpointListener(
 				)
 			}
 			else
-				send(await execute(requestish))
+				send(await endpoint(requestish))
 		}
 		catch (error) {
 			res.statusCode = 500