+++ s/transports/http/server.ts
@@ -1,12 +1,13 @@
 
 import {Server} from "http"
-import {Endpoint} from "../../core/types.js"
+import {defaults} from "../defaults.js"
+import {Endpoint, ServerMeta} from "../../core/types.js"
 import {allowCors} from "./node-utils/listener-transforms/allow-cors.js"
 import {healthCheck} from "./node-utils/listener-transforms/health-check.js"
 import {EndpointListenerOptions, makeEndpointListener} from "./node-utils/endpoint-listener.js"
 
 export class HttpServer extends Server {
-	constructor(endpoint: Endpoint, options: EndpointListenerOptions = {}) {
+	constructor(endpoint: (meta: ServerMeta) => Endpoint, options: EndpointListenerOptions = {}) {
 		super(
 			allowCors(
 				healthCheck(
@@ -15,6 +16,7 @@ export class HttpServer extends Server {
 				)
 			)
 		)
+		this.timeout = options.timeout ?? defaults.timeout
 	}
 }