+++ s/transports/websocket/client.ts
@@ -33,14 +33,14 @@ export async function webSocketRemote<F extends Fns>(
 		const socketry = new Socketry({
 			socket,
 			timeout,
-			headers: {},
 			onError,
 		})
 
 		const fns = remote<F>(socketry.remoteEndpoint, params)
 
-		socket.onmessage = socketry.prepareMessageHandler(
-			getLocalEndpoint(fns as F)
+		socket.onmessage = event => socketry.receive(
+			getLocalEndpoint(fns as F),
+			event,
 		)
 
 		await ready