+++ s/transports/websocket/examples/apis.ts
@@ -1,6 +1,4 @@
 
-import {api} from "../../../core/types.js"
-
 export type ExampleServersideFns = {
 	now(): Promise<number>
 }
@@ -9,24 +7,21 @@ export type ExampleClientsideFns = {
 	sum(a: number, b: number): Promise<number>
 }
 
-export const exampleServersideApi = (
-		clientside: ExampleClientsideFns,
-	) => api((): ExampleServersideFns => ({
+export const exampleServersideApi =
+	(clientside: ExampleClientsideFns): ExampleServersideFns => ({
 
 	async now() {
 		await clientside.sum(1, 2)
 		return Date.now()
 	},
-}))
+})
 
-export const exampleClientsideApi = (
-		_serverside: ExampleServersideFns,
-		rememberCall: () => void
-	) => api(() => ({
+export const exampleClientsideApi =
+	(_serverside: ExampleServersideFns, rememberCall: () => void) => ({
 
 	async sum(a: number, b: number) {
 		rememberCall()
 		return a + b
 	},
-}))
+})