+++ s/transports/websocket/examples/server.ts
@@ -1,15 +1,15 @@
 
 import {WebSocketServer} from "../server.js"
 import {remote} from "../../../core/remote.js"
-import {expose} from "../../../core/expose.js"
+import {endpoint} from "../../../core/endpoint.js"
 import {ExampleClientsideFns, exampleServersideApi} from "./apis.js"
 
 const server = new WebSocketServer({
-	acceptConnection: ({remoteEndpoint}) => {
+	acceptConnection: async({remoteEndpoint}) => {
 		const clientside = remote<ExampleClientsideFns>(remoteEndpoint)
 		return {
 			closed: () => {},
-			localEndpoint: expose(exampleServersideApi(clientside)),
+			localEndpoint: endpoint(exampleServersideApi(clientside)),
 		}
 	},
 })